src/main/java/com/jaenyeong/dev/springboot/web/ProfileController.java
package com.jaenyeong.dev.springboot.web;

import lombok.RequiredArgsConstructor;
import org.springframework.core.env.Environment;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.Arrays;
import java.util.List;

@RequiredArgsConstructor
@RestController
public class ProfileController {
    private final Environment environment;

    @GetMapping("/profile")
    public String profile() {
        /**
         * environment.getActiveProfiles();
         * 현재 실행중인 Active Profile을 다 가져옴 (여기서는 real, oauth, real-db 세개)
         * real은 step2를 다시 사용할 수 있어 남겨둠
         * 무중단 배포는 real1, real2 사용
         */
        List<String> profiles = Arrays.asList(environment.getActiveProfiles());
        List<String> realProfiles = Arrays.asList("real", "real1", "real2");

        String defaultProfile = profiles.isEmpty() ? "default" : profiles.get(0);

        return profiles.stream()
                .filter(realProfiles::contains)
                .findAny()
                .orElse(defaultProfile);
    }

}
